Credi-cart-pay

+++ Controllers/OrderController.cs
@@ -1,6 +1,9 @@
 using System.Threading.Tasks;
 using dotnet_store.Models;
 using dotnet_store.Services;
+using Iyzipay;
+using Iyzipay.Model;
+using Iyzipay.Request;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
@@ -12,11 +15,13 @@ public class OrderController : Controller
 {
     private ICartService _cartService;
     private readonly DataContext _context;
+    private readonly IConfiguration _configuration;
 
-    public OrderController(ICartService cartService, DataContext context)
+    public OrderController(ICartService cartService, DataContext context, IConfiguration configuration)
     {
         _cartService = cartService;
         _context = context;
+        _configuration = configuration;
     }
 
     [Authorize(Roles = "Admin")]
@@ -60,11 +65,12 @@ public class OrderController : Controller
                 Telefon = model.Telefon,
                 AdresSatiri = model.AdresSatiri,
                 PostaKodu = model.PostaKodu,
+                SiparisTarihi = DateTime.Now,
                 Sehir = model.Sehir,
                 SiparisNotu = model.SiparisNotu!,
                 ToplamFiyat = cart.Toplam(),
                 UserName = username,
-                OrderItems = cart.CartItems.Select(ci => new OrderItem
+                OrderItems = cart.CartItems.Select(ci => new Models.OrderItem
                 {
                     UrunId = ci.UrunId,
                     Fiyat = ci.Urun.Fiyat,
@@ -72,12 +78,17 @@ public class OrderController : Controller
                 }).ToList()
             };
 
-            _context.Orders.Add(order);
-            _context.Carts.Remove(cart);
+            var payment = await ProcessPayment();
 
-            await _context.SaveChangesAsync();
+            if (payment.Status == "success")
+            {
+                _context.Orders.Add(order);
+                _context.Carts.Remove(cart);
 
-            return RedirectToAction("Completed", new { orderId = order.Id });
+                await _context.SaveChangesAsync();
+
+                return RedirectToAction("Completed", new { orderId = order.Id });
+            }
         }
 
         ViewBag.Cart = cart;
@@ -89,9 +100,107 @@ public class OrderController : Controller
         return View("Completed", orderId);
     }
 
-    public ActionResult OrderList()
+    public async Task<ActionResult> OrderList()
     {
-        return View();
+        var username = User.Identity?.Name;
+
+        var PersonProductList = await _context.Orders
+                                        .Include(i => i.OrderItems)
+                                        .ThenInclude(i => i.Urun)
+                                        .Where(i => i.UserName == username)
+                                        .ToListAsync();
+        return View(PersonProductList);
+    }
+
+    private async Task<Payment> ProcessPayment()
+    {
+        Options options = new Options();
+        options.ApiKey = _configuration["PaymentAPI:APIKey"];
+        options.SecretKey = _configuration["PaymentAPI:SecretKey"];
+        options.BaseUrl = "https://sandbox-api.iyzipay.com";
+
+        CreatePaymentRequest request = new CreatePaymentRequest();
+        request.Locale = Locale.TR.ToString();
+        request.ConversationId = "123456789";
+        request.Price = "1";
+        request.PaidPrice = "1.2";
+        request.Currency = Currency.TRY.ToString();
+        request.Installment = 1;
+        request.BasketId = "B67832";
+        request.PaymentChannel = PaymentChannel.WEB.ToString();
+        request.PaymentGroup = PaymentGroup.PRODUCT.ToString();
+
+        PaymentCard paymentCard = new PaymentCard();
+        paymentCard.CardHolderName = "John Doe";
+        paymentCard.CardNumber = "5528790000000008";
+        paymentCard.ExpireMonth = "12";
+        paymentCard.ExpireYear = "2030";
+        paymentCard.Cvc = "123";
+        paymentCard.RegisterCard = 0;
+        request.PaymentCard = paymentCard;
+
+        Buyer buyer = new Buyer();
+        buyer.Id = "BY789";
+        buyer.Name = "John";
+        buyer.Surname = "Doe";
+        buyer.GsmNumber = "+905350000000";
+        buyer.Email = "email@email.com";
+        buyer.IdentityNumber = "74300864791";
+        buyer.LastLoginDate = "2015-10-05 12:43:35";
+        buyer.RegistrationDate = "2013-04-21 15:12:09";
+        buyer.RegistrationAddress = "Nidakule Göztepe, Merdivenköy Mah. Bora Sok. No:1";
+        buyer.Ip = "85.34.78.112";
+        buyer.City = "Istanbul";
+        buyer.Country = "Turkey";
+        buyer.ZipCode = "34732";
+        request.Buyer = buyer;
+
+        Address shippingAddress = new Address();
+        shippingAddress.ContactName = "Jane Doe";
+        shippingAddress.City = "Istanbul";
+        shippingAddress.Country = "Turkey";
+        shippingAddress.Description = "Nidakule Göztepe, Merdivenköy Mah. Bora Sok. No:1";
+        shippingAddress.ZipCode = "34742";
+        request.ShippingAddress = shippingAddress;
+
+        Address billingAddress = new Address();
+        billingAddress.ContactName = "Jane Doe";
+        billingAddress.City = "Istanbul";
+        billingAddress.Country = "Turkey";
+        billingAddress.Description = "Nidakule Göztepe, Merdivenköy Mah. Bora Sok. No:1";
+        billingAddress.ZipCode = "34742";
+        request.BillingAddress = billingAddress;
+
+        List<BasketItem> basketItems = new List<BasketItem>();
+        BasketItem firstBasketItem = new BasketItem();
+        firstBasketItem.Id = "BI101";
+        firstBasketItem.Name = "Binocular";
+        firstBasketItem.Category1 = "Collectibles";
+        firstBasketItem.Category2 = "Accessories";
+        firstBasketItem.ItemType = BasketItemType.PHYSICAL.ToString();
+        firstBasketItem.Price = "0.3";
+        basketItems.Add(firstBasketItem);
+
+        BasketItem secondBasketItem = new BasketItem();
+        secondBasketItem.Id = "BI102";
+        secondBasketItem.Name = "Game code";
+        secondBasketItem.Category1 = "Game";
+        secondBasketItem.Category2 = "Online Game Items";
+        secondBasketItem.ItemType = BasketItemType.VIRTUAL.ToString();
+        secondBasketItem.Price = "0.5";
+        basketItems.Add(secondBasketItem);
+
+        BasketItem thirdBasketItem = new BasketItem();
+        thirdBasketItem.Id = "BI103";
+        thirdBasketItem.Name = "Usb";
+        thirdBasketItem.Category1 = "Electronics";
+        thirdBasketItem.Category2 = "Usb / Cable";
+        thirdBasketItem.ItemType = BasketItemType.PHYSICAL.ToString();
+        thirdBasketItem.Price = "0.2";
+        basketItems.Add(thirdBasketItem);
+        request.BasketItems = basketItems;
+
+        return await Payment.Create(request, options);
     }
 
 }
